feat(optional graph cycler): started work on optional statistics cycle
